implemented AgsAutomationEditor to use GUI scale factor

+++ ags/X/editor/ags_notation_edit_callbacks.c
@@ -843,7 +843,7 @@ ags_notation_edit_hscrollbar_value_changed(GtkRange *range, AgsNotationEdit *not
     g_free(str);
   }
 
-  value = GTK_RANGE(notation_edit->hscrollbar)->adjustment->value / (gui_scale_factor * 64.0);
+  value = GTK_RANGE(notation_edit->hscrollbar)->adjustment->value / (guint) (gui_scale_factor * 64.0);
   gtk_adjustment_set_value(notation_edit->ruler->adjustment,
 			   value);
   gtk_widget_queue_draw((GtkWidget *) notation_edit->ruler);
